backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
pydantic[email]==2.5.0
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3
pdfplumber==0.9.0
opencv-python==4.8.1.78
numpy==1.24.3
boto3==1.34.0
python-magic==0.4.27
aiofiles==23.2.0
